requirements.txt
asgiref==3.2.5
certifi==2019.11.28
Django==3.0.4
django-cors-headers==3.2.1
django-environ==0.4.5
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
numpy==1.18.1
pandas==1.0.2
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
wincertstore==0.2
